Fix all warnings...
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@734 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/file_chunk.cpp
@@ -54,10 +54,10 @@ FileChunk::~FileChunk()
 void FileChunk::Merge(FileChunk chunk)
 {
 	/* Check we are overlapping */
-	if(offset + size < chunk.GetOffset())
+	if(offset + (off_t)size < chunk.GetOffset())
 		return;
 
-	if(offset > chunk.GetOffset() + chunk.GetSize())
+	if(offset > chunk.GetOffset() + (off_t)chunk.GetSize())
 		return;
 
 	/* Merge */
